Goldstein-Price Function (F18)

A multimodal test function with four local minima and one global minimum. Fixed dimension of 2.

F18(x)

Arguments

x

Numeric vector of length 2 (2-dimensional input).

Value

Numeric scalar representing the function value.

Details

Formula: $$f(x) = (1 + (x_1 + x_2 + 1)^2(19 - 14x_1 + 3x_1^2 - 14x_2 + 6x_1 x_2 + 3x_2^2)) \times (30 + (2x_1 - 3x_2)^2(18 - 32x_1 + 12x_1^2 + 48x_2 - 36x_1 x_2 + 27x_2^2))$$

Global minimum: \(f(0, -1) = 3\)

Characteristics:

Type: Multimodal

Separable: No

Differentiable: Yes

Fixed dimension: 2

Number of local minima: 4

Default bounds: \([-2, 2]^2\)

The Goldstein-Price function has a complex landscape with the global minimum surrounded by local minima of increasing value.

See also

Examples

F18(c(0, -1))  # Returns 3 (global minimum)
#> [1] 3
F18(c(0, 0))   # Returns 600
#> [1] 600